requirements.txt
ipython==8.12.3
langchain_community==0.3.0
langchain_core==0.3.5
langchain_experimental==0.3.1
langchain_openai==0.2.0
langgraph==0.2.27
matplotlib==3.9.2
numpy==2.1.1
python-dotenv==1.0.1
